// src/app/interfaces/get-card-available-raffle-response.ts
export interface GetCardAvailableRaffleResponse {

  message: string;
    success: boolean;
    error: boolean;
    code: string;
    date: string;
    data: {
        Card: PaginatedCards;
    };
}


export interface PaginatedCards {
    current_page: number;
    data: Card[];
    first_page_url: string;
    from: number;
    last_page: number;
    last_page_url: string;
    links: PaginationLink[];
    next_page_url: string | null;
    path: string;
    per_page: number;
    prev_page_url: string | null;
    to: number;
    total: number;
}

export interface Card {
    id: number;
    name: string | null;
    pos01: number;
    pos02: number;
    pos03: number;
    pos04: number;
    pos05: number;
    pos06: number;
    pos07: number;
    pos08: number;
    pos09: number;
    pos10: number;
    pos11: number;
    pos12: number;
    pos13: number;
    pos14: number;
    pos15: number;
    pos16: number;
    pos17: number;
    pos18: number;
    pos19: number;
    pos20: number;
    pos21: number;
    pos22: number;
    pos23: number;
    pos24: number;
    pos25: number;
    comb01: string;
    comb02: string;
    comb03: string;
    comb04: string;
    comb05: string;
    comb06: string;
    comb07: string;
    comb08: string;
    comb09: string;
    comb10: string;
    comb11: string;
    comb12: string;
    comb13: string;
    comb14: string;
    combTotal: string;
    desc_pos01: string;
    desc_pos02: string;
    desc_pos03: string;
    desc_pos04: string;
    desc_pos05: string;
    desc_pos06: string;
    desc_pos07: string;
    desc_pos08: string;
    desc_pos09: string;
    desc_pos10: string;
    desc_pos11: string;
    desc_pos12: string;
    desc_pos13: string;
    desc_pos14: string;
    desc_pos15: string;
    desc_pos16: string;
    desc_pos17: string;
    desc_pos18: string;
    desc_pos19: string;
    desc_pos20: string;
    desc_pos21: string;
    desc_pos22: string;
    desc_pos23: string;
    desc_pos24: string;
    desc_pos25: string;
    desc_comb01: string;
    desc_comb02: string;
    desc_comb03: string;
    desc_comb04: string;
    desc_comb05: string;
    desc_comb06: string;
    desc_comb07: string;
    desc_comb08: string;
    desc_comb09: string;
    desc_comb10: string;
    desc_comb11: string;
    desc_comb12: string;
    desc_comb13: string;
    desc_comb14: string;
    desc_combTotal: string;
    active: string;
    groupFicha_id: number | null;
    start_date: Date;
    end_date: Date | null;
    created_at: Date | null;
    updated_at: Date | null;
    card_raffle_id?: number|null;
}

export interface PaginationLink {
    url: string | null;
    label: string;
    active: boolean;
}